ASoC: msm: Add support timestamp update VCAP playback usecase
Compressed data in HDMI IN or VCAP usecase is propagated with
timestamp from DSP with each buffer. For playback, the same or
interpolated timestamp is updated with the corresponding buffer.
DSP will render the buffer a) without any delay if timestamp mode is
synchronous to absolute timestamp b) with a delay if timestamp is more
than the absolute timestamp c) drop the buffer if the timestamp is
less than the absolute timestamp. Add support for the same.

Signed-off-by: Krishnankutty Kolathappilly <kkolat@codeaurora.org>
(cherry picked from commit d04445f5d1fa86d94e98918c2290fc79fdd9c6ad)

Change-Id: I155056aa63b1937163dda2b71ddc209fa3a1abed
Signed-off-by: Sivasri Kumar Vanka <sivasri@codeaurora.org>